Add flutter_redux and implement in memory authentication service
The app is functional now. However, the navigation not finished.
Rather than navigating from within the widgets, I would like to
be able to dispatch actions that will cause redux to manage the
navigation. This way, navigation could occur based on application
state, rather than on only user interactions. But this will have
to wait for a future coding session.